mensagem de confirmação ao excluir

+++ src/view/listarNoticias.php
@@ -56,7 +56,7 @@
                     <td><?= $noticias[$i][2] ?></td>
                     <td><?= $noticias[$i][4] ?></td>
                     <td><a class="btn btn-warning" href="?section=NoticiaControle&function=detalharNoticia&id=<?= $noticias[$i][0] ?>">Detalhar</a></td>
-                    <td><a class="btn btn-danger" href="?section=NoticiaControle&function=excluirNoticia&id=<?= $noticias[$i][0] ?>">Excluir</a></td>
+                    <td><a class="btn btn-danger" onclick="return  confirm('Tem certeza que deseja apagar esta notícia?')" href="?section=NoticiaControle&function=excluirNoticia&id=<?= $noticias[$i][0] ?>">Excluir</a></td>
                 </tr>
             <?php endfor;
         } ?>
